+++ packages/sources/svelte/rollup.config.js
@@ -16,7 +16,7 @@ const src = {
     },
     {
       folder: 'indicators',
-      components: ['VtmnBadge', 'VtmnPrice', 'VtmnTag'],
+      components: ['VtmnBadge', 'VtmnPrice', 'VtmnRating', 'VtmnTag'],
     },
     {
       folder: 'navigation',
@@ -24,7 +24,13 @@ const src = {
     },
     {
       folder: 'overlays',
-      components: ['VtmnPopover'],
+      components: [
+        'VtmnAlert',
+        'VtmnModal',
+        'VtmnPopover',
+        'VtmnSnackbar',
+        'VtmnToast',
+      ],
     },
     {
       folder: 'selection-controls',
@@ -32,7 +38,13 @@ const src = {
     },
     {
       folder: 'structure',
-      components: ['VtmnAccordion', 'VtmnDivider', 'VtmnList', 'VtmnListItem'],
+      components: [
+        'VtmnAccordion',
+        'VtmnCard',
+        'VtmnDivider',
+        'VtmnList',
+        'VtmnListItem',
+      ],
     },
   ],
   guidelines: [