+++ packages/sources/svelte/src/components/overlays/VtmnAlert/enums.js
@@ -1,0 +1,8 @@
+export const VTMN_ALERT_TIMEOUT = 8000;
+
+export const VTMN_ALERT_VARIANT = {
+  INFO: 'info',
+  WARNING: 'warning',
+  SUCCESS: 'success',
+  DANGER: 'danger',
+};