+++ packages/sources/svelte/src/components/structure/VtmnCard/enums.js
@@ -1,0 +1,5 @@
+export const VTMN_CARD_VARIANT = {
+  TOP_IMAGE: 'top-image',
+  SIDE_IMAGE: 'side-image',
+  FULL_IMAGE: 'full-image',
+};